readline: Skip the test_interrupt_in_other_thread on arm32.
Skip the test_interrupt_in_other_thread test failing on Travis arm32.
The failing test is not a new issue. It had already been skipped in .travis.yml
before updating .travis.yml.
https://github.com/ruby/ruby/blob/7b05cb8dbbd637910757f402e64da3908b4bb809/.travis.yml#L99-L101

Note that RUBY_PLATFORM is "armv8l-linux-eabihf" on Travis arm32 pipeline.
And the readline was deleted in both master and ruby_3_3 branches.
https://github.com/ruby/ruby/commit/59fd67fc3d405e529e038172e769ff20a8fb5535

So, we only need this commit on ruby_3_2 and older Ruby version branches.