Improve method_missing method documentation
This enhances the documentation of method_missing by documenting some
additional behavior.

First it enhances the description to make it clear you can also pass in
kwargs and a block.

Then it simplifies the example by relying on a Ruby 3.0 feature to pass
in ... instead of individual arguments. The super call without braces
and arguments means it inherits them all, regardless of method
definition.