s/SafeStringValue/StringValue/
The macro SafeStringValue() became just StringValue() in c5c05460ac2,
and it is deprecated nowadays.

This patch replaces remaining macro usage. Some occurrences are left in
ext/stringio and ext/win32ole, they should be fixed upstream.

The macro itself is not deleted, because it may be used in extensions.