Make Net::HTTP tests ensure local & remote hosts are same addr family
* When making connections wiht TCPSocket.new and friends, we resolve the
  passed-in hostname with AI_ADDRCONFIG.
* AI_ADDRCONFIG means (in practice) "do not make AAAA queries or return
  any IPv6 addresses unless this machine has a network interface with an
  IPv6 address".
* The loopback adapter does _NOT_ count as "a network interface with an
  IPv6 address" for this check; your machine must have some _other_, non
  loopback address.

This causes issues with the Net::HTTP tests which set the local_host to
connect _from_. This obviously needs to have the same address family as
the host to connect _to_. However, we call Addrinfo.tcp('localhost')
 _without_ AI_ADDRCONFIG to set this, which means it might be `::1` even
in circumstances where the lookup done on the hostname to connect _to_
(also 'localhost') returns `127.0.0.1`.

Fix this by passing a hostname into `local_host=`, not an address; the
hostname will be resolved with the same address family used for
connecting.

[Bug #20208]