Only execute WBs when we really have to
This patch adds code to YJIT to avoid executing write barriers unless we
really must.  We must execute the write barrier in either of these
conditions:

1. We are writing a reference to an old object
2. The GC is performing incremental marking

We can test that the object is old by checking the age bits in the
object header.  If the object is old, we must run the WB.  If the GC is
performing incremental marking, it will call a callback letting YJIT
know that we must always execute the WB during that time.  Our callback
sets a global, then we test the value in the global.