[ruby/prism] Speed up creating Ruby AST
When creating the Ruby AST, we were previously allocating Location
objects for every node and every inner location. Instead, this
commit changes it to pack both the start offset and length into a
single u64 and pass that into the nodes. Then, when the locations
are requested via a reader method, we lazily allocate the Location
objects.

https://github.com/ruby/prism/commit/de203dca83

Co-Authored-By: Aaron Patterson <tenderlove@ruby-lang.org>