[rubygems/rubygems] Get rid of too long paths
This is a fix-up of #7483.

Test method names in Rubygems are often very long, and the path under
the temporary directory generated from them can easily exceed system
limits.

Even without including the method name in the path name, by saving the
method and path name pair, it is possible to find the method from the
remaining path name.

https://github.com/rubygems/rubygems/commit/de55064a3d