Module#remove_const: cleanup vm->defined_module_hash
[Bug #20311]

All module and classes defined from C are stored in this table
to ensure they are pinned and won't be moved by the compactor.

However this means we need to remove them from it whenever a constant
referencing them is removed.

But since a same class or module may be referenced under multiple
names, we have to do some ref counting.