Avoid a hash allocation when calling def f(kw: 1) with keyword splat
c20e819e8b04e84a4103ca9a036022543459c213 made calling a method
that accepts specific keywords but not a keyword splat with a keyword
splat allocating a hash when it previously did not.

Instead of duplicating the hash and removing hash entries,
lookup in the provided hash without duplicating the hash, and
track to make sure all hash entries in the hash were looked up
(extra hash entries would be errors).